Tweak data node request index handling (#118542)
Small tweak around how data node requests handle no indices w.r.t.
shards.
